refactor(protocol): merge fragment modules into their owning files
- policy/version.ts (1 LOC const) -> policy/definition.ts
- policy/plan.ts (single zod object) -> inlined in policy/index.ts
- app-connector/schema-primitives.ts -> app-connector/definition.ts
- app-connector/schemas.ts (pure re-export duplicate) deleted; index.ts
  repointed to definition.js/installation.js
