--- packages/protocol/src/app-connector/schemas.ts
@@ -1,2 +1,0 @@
-export * from "./definition.js";
-export * from "./installation.js";
