--- packages/protocol/src/policy/version.ts
@@ -1,1 +1,0 @@
-export const policyKernelVersion = 1;
